Refactor: 이미지 업로드 기능을 misc로 빼기
