+++ app-server/subprojects/bounded_context/accessibility/infra/build.gradle.kts
@@ -9,6 +9,7 @@ dependencies {
     implementation(projects.boundedContext.challenge.domain)
     implementation(projects.boundedContext.challenge.application)
     implementation(projects.boundedContext.challenge.infra)
+    implementation(projects.boundedContext.misc.application)
     implementation(projects.crossCuttingConcern.infra.network)
     implementation(projects.crossCuttingConcern.infra.persistenceModel)
     implementation("org.springframework.boot:spring-boot-starter-web")
