+++ app-server/subprojects/bounded_context/misc/application/src/main/kotlin/club/staircrusher/image/application/port/in/use_case/GetImageUploadUrlsUseCase.kt
@@ -1,7 +1,8 @@
-package club.staircrusher.accessibility.application.port.`in`
+package club.staircrusher.image.application.port.`in`.use_case
 
-import club.staircrusher.accessibility.application.port.out.file_management.FileManagementService
-import club.staircrusher.accessibility.application.port.out.file_management.UploadUrl
+import club.staircrusher.image.application.port.out.file_management.FileManagementService
+import club.staircrusher.image.application.port.out.file_management.ImageUploadPurposeType
+import club.staircrusher.image.application.port.out.file_management.UploadUrl
 import club.staircrusher.stdlib.di.annotation.Component
 
 @Component
@@ -11,6 +12,7 @@ class GetImageUploadUrlsUseCase(
     fun handle(
         urlCount: Int,
         filenameExtension: String,
+        imageUploadPurposeType: ImageUploadPurposeType,
     ): List<UploadUrl> {
         if (urlCount > urlCountLimit) {
             throw IllegalArgumentException("한 번에 최대 ${urlCountLimit}개의 URL만 생성할 수 있습니다.")
@@ -20,7 +22,7 @@ class GetImageUploadUrlsUseCase(
                 "$allowedFilenameExtensions 중 하나의 확장자를 지정해주세요.")
         }
         return (0 until urlCount).map {
-            fileManagementService.getFileUploadUrl(filenameExtension)
+            fileManagementService.getFileUploadUrl(filenameExtension, imageUploadPurposeType)
         }
     }
