+++ app-server/subprojects/cross_cutting_concern/test/spring_it/src/main/kotlin/club/staircrusher/testing/spring_it/mock/MockFileManagementService.kt
@@ -1,7 +1,8 @@
 package club.staircrusher.testing.spring_it.mock
 
-import club.staircrusher.accessibility.application.port.out.file_management.FileManagementService
-import club.staircrusher.accessibility.application.port.out.file_management.UploadUrl
+import club.staircrusher.image.application.port.out.file_management.FileManagementService
+import club.staircrusher.image.application.port.out.file_management.ImageUploadPurposeType
+import club.staircrusher.image.application.port.out.file_management.UploadUrl
 import club.staircrusher.stdlib.clock.SccClock
 import org.springframework.core.io.ClassPathResource
 import java.io.ByteArrayOutputStream
@@ -9,7 +10,7 @@ import java.io.File
 import java.nio.file.Path
 
 class MockFileManagementService : FileManagementService {
-    override fun getFileUploadUrl(fileExtension: String): UploadUrl {
+    override fun getFileUploadUrl(fileExtension: String, purposeType: ImageUploadPurposeType): UploadUrl {
         return UploadUrl(
             url = "example.$fileExtension",
             expireAt = SccClock.instant().plusSeconds(60L)
@@ -20,7 +21,7 @@ class MockFileManagementService : FileManagementService {
         return ClassPathResource("example.png").file
     }
 
-    override suspend fun uploadImage(fileName: String, fileBytes: ByteArray): String {
+    override suspend fun uploadAccessibilityImage(fileName: String, fileBytes: ByteArray): String {
         return fileName
     }
